BIOS 6640  HW Suggestions for Documentation
The documentation on how to download and start using the package is very clear. I am a little confused on the navigation of the repository though. I can figure out the contents of some folders just by looking at the files, like the R, data, examples, and references folders, and some of the commit messages, but there are others that I'm less clear about like the man and src folders. Is that something that is usually done when creating a repository? Hope my suggestion is helpful!